refactor: MVC패턴을 적용했기에 변경된 파일 구조를 TEST에 반영
