+++ src/test/java/racingcar/Input/CorrectInputTest.java
@@ -3,7 +3,7 @@ package racingcar.Input;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static racingcar.view.InputView.isCorrectInput;
+import static racingcar.service.InputService.isCorrectInput;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
